[posixlibchange hardcoded poll constants to imported (#4626)

+++ javalib/src/main/scala/java/net/UnixPlainDatagramSocketImpl.scala
@@ -49,7 +49,7 @@ private[net] class UnixPlainDatagramSocketImpl
 
     pollFd.fd = fd.fd
     pollFd.revents = 0
-    pollFd.events = POLLIN
+    pollFd.events = POLLIN.toShort
 
     val pollRes = poll(pollFd, nAlloc, timeout)
     val revents = pollFd.revents
